Switch from CodeMirror to AceEditor
Seems to have better documentation, easier to work with.
This commit is the first time I have references working
properly for the editor.

This commit also modifies the build process for the frontend
a little bit. Dockerfile, package.json, and package-lock.json
have been updated with everything needed for this editor.