Add New and Delete button functionality
All backbone changes necessary for New and Delete button functionalities.
Not perfect yet, and the frontend is messy, but the backbone of the program is more
or less complete now.
Everything is communicating with everything and all that. Next step is data persistence
on the server, then writing the interpreter itself, then cleaning up the frontend to mak e
it halfway decent and useable.