Reword frontend to use file index instead of file object
Did not want to have to making sure the references to the selected file
are always the same as the actual file in the list of files. Decided
to just make it an index variable, and we will access the list whenever
needed so that references don't need to be worried about.

Also, while it should have been a separate commit, this commit will
also introduce a new server request on the frotend side, as I have decided
to do the processing of the code on the server side.